refactor: FSD Phase 6h-2 — AI チャットのフックと MessageInput を pages/ask-ai へ移し hooks/components を撤去 (FRESTYLE-175)
AI チャット(ask-ai 画面)専用のフックと入力コンポーネントを pages/ask-ai へ
集約する。

- useAskAi / useAiChat / useAiSession / useAiChatSse → pages/ask-ai/model
- MessageInput → pages/ask-ai/ui

useAskAi が他 3 フックを束ね、AskAiPage だけが useAskAi と MessageInput を
使う。すべて ask-ai 画面に閉じているため相対参照(barrel なし)。

これで旧 hooks/ と components/ ディレクトリが空になり撤去。src 直下の
非 FSD ディレクトリ(components / hooks / constants / utils / lib / store /
types)がすべて消え、app / pages / widgets / features / entities / shared の
6 レイヤーに揃った(assets / generated / test を除く)。

既存テストの assert は変更していない。テスト件数 1184 で不変。
